Move ReportsController helper methods to ReportsHelper
