+++ app/helpers/reports_helper.rb
@@ -1,0 +1,27 @@
+module ReportsHelper
+  def has_unread?(post)
+    return false unless @opened_posts.present?
+    view = @opened_posts.detect { |v| v.post_id == post.id }
+    return false unless view
+    return false if view.ignored?
+    return false if view.read_at.nil? # totally unread, not partially
+    view.read_at < post.tagged_at
+  end
+
+  def never_read?(post)
+    return false unless logged_in?
+    return true unless @opened_posts.present?
+    view = @opened_posts.detect { |v| v.post_id == post.id }
+    return true unless view
+    return false if view.ignored?
+    view.read_at.nil?
+  end
+
+  def ignored?(post)
+    return false unless @opened_posts.present?
+    view = @opened_posts.detect { |v| v.post_id == post.id }
+    board_view = @board_views.detect { |v| v.board_id == post.board_id }
+    return false unless view || board_view
+    view.try(:ignored?) || board_view.try(:ignored?)
+  end
+end
